requirements.txt
playwright>=1.40.0
python-dotenv>=1.0.0
supabase>=2.0.0
